Choose a site to publish to in a modal, sharing onboarding's connect step
