+++ apps/ui/src/components/site-overview-view/connections-tab.module.css
@@ -79,10 +79,6 @@
 	margin-block-start: var(--wpds-dimension-padding-md);
 }
 
-.picker {
-	max-width: 420px;
-}
-
 /* Icon SVGs carry no fill of their own; without this they paint black in both
    colour schemes. */
 .rowActions svg,
